reorg code structure
Move backend implementation into a separate mod so that this
looks more like a library.
